Add layer reducer
