+++ packages/noya-state/src/reducers/applicationReducer.ts
@@ -8,12 +8,10 @@ import * as Layers from '../layers';
 import {
   findPageLayerIndexPaths,
   getCurrentComponentsTab,
-  getCurrentPage,
   getCurrentPageIndex,
   getCurrentTab,
   getSelectedLayerIndexPaths,
 } from '../selectors/selectors';
-import { SelectionType, updateSelection } from '../utils/selection';
 import { AlignmentAction, alignmentReducer } from './alignmentReducer';
 import { CanvasAction, canvasReducer } from './canvasReducer';
 import {
@@ -25,6 +23,7 @@ import {
   LayerPropertyAction,
   layerPropertyReducer,
 } from './layerPropertyReducer';
+import { LayerAction, layerReducer } from './layerReducer';
 import { PageAction, pageReducer } from './pageReducer';
 import {
   StringAttributeAction,
@@ -58,12 +57,7 @@ export type Action =
   | [type: 'setTab', value: WorkspaceTab]
   | PageAction
   | CanvasAction
-  | [type: 'deleteLayer', layerId: string | string[]]
-  | [
-      type: 'selectLayer',
-      layerId: string | string[] | undefined,
-      selectionType?: SelectionType,
-    ]
+  | LayerAction
   | StyleAction
   | StringAttributeAction
   | AlignmentAction
@@ -105,39 +99,9 @@ export function applicationReducer(
     case 'setFixedRadius':
     case 'setLayerRotation':
       return layerPropertyReducer(state, action);
-    case 'deleteLayer': {
-      const [, id] = action;
-
-      const ids = typeof id === 'string' ? [id] : id;
-
-      const page = getCurrentPage(state);
-      const pageIndex = getCurrentPageIndex(state);
-      const indexPaths = Layers.findAllIndexPaths(page, (layer) =>
-        ids.includes(layer.do_objectID),
-      );
-
-      // We delete in reverse so that the indexPaths remain accurate even
-      // after some layers are deleted.
-      const reversed = [...indexPaths.reverse()];
-
-      return produce(state, (draft) => {
-        reversed.forEach((indexPath) => {
-          const childIndex = indexPath[indexPath.length - 1];
-          const parent = Layers.access(
-            draft.sketch.pages[pageIndex],
-            indexPath.slice(0, -1),
-          ) as Layers.ParentLayer;
-          parent.layers.splice(childIndex, 1);
-        });
-      });
-    }
-    case 'selectLayer': {
-      const [, id, selectionType = 'replace'] = action;
-
-      return produce(state, (draft) => {
-        updateSelection(draft.selectedObjects, id, selectionType);
-      });
-    }
+    case 'deleteLayer':
+    case 'selectLayer':
+      return layerReducer(state, action);
     case 'distributeLayers':
     case 'alignLayers': {
       return alignmentReducer(state, action);
