+++ packages/noya-state/src/reducers/layerReducer.ts
@@ -1,0 +1,56 @@
+import produce from 'immer';
+import * as Layers from '../layers';
+import { getCurrentPage, getCurrentPageIndex } from '../selectors/selectors';
+import { SelectionType, updateSelection } from '../utils/selection';
+import { ApplicationState } from './applicationReducer';
+
+export type LayerAction =
+  | [type: 'deleteLayer', layerId: string | string[]]
+  | [
+      type: 'selectLayer',
+      layerId: string | string[] | undefined,
+      selectionType?: SelectionType,
+    ];
+
+export function layerReducer(
+  state: ApplicationState,
+  action: LayerAction,
+): ApplicationState {
+  switch (action[0]) {
+    case 'deleteLayer': {
+      const [, id] = action;
+
+      const ids = typeof id === 'string' ? [id] : id;
+
+      const page = getCurrentPage(state);
+      const pageIndex = getCurrentPageIndex(state);
+      const indexPaths = Layers.findAllIndexPaths(page, (layer) =>
+        ids.includes(layer.do_objectID),
+      );
+
+      // We delete in reverse so that the indexPaths remain accurate even
+      // after some layers are deleted.
+      const reversed = [...indexPaths.reverse()];
+
+      return produce(state, (draft) => {
+        reversed.forEach((indexPath) => {
+          const childIndex = indexPath[indexPath.length - 1];
+          const parent = Layers.access(
+            draft.sketch.pages[pageIndex],
+            indexPath.slice(0, -1),
+          ) as Layers.ParentLayer;
+          parent.layers.splice(childIndex, 1);
+        });
+      });
+    }
+    case 'selectLayer': {
+      const [, id, selectionType = 'replace'] = action;
+
+      return produce(state, (draft) => {
+        updateSelection(draft.selectedObjects, id, selectionType);
+      });
+    }
+    default:
+      return state;
+  }
+}
